ADD config to jwt auth
