+++ backend/interfaces/api/server/auth/config.go
@@ -1,0 +1,10 @@
+package auth
+
+import (
+	"github.com/yoshihiro-shu/draft-backend/internal/config"
+)
+
+type authConfig struct {
+	AccessToken  config.AuthToken
+	RefreshToken config.AuthToken
+}
