+++ backend/interfaces/api/server/auth/jwt.go
@@ -2,28 +2,36 @@ package auth
 
 import (
 	"fmt"
-	"net/http"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/dgrijalva/jwt-go"
+	"github.com/yoshihiro-shu/draft-backend/internal/config"
 )
 
 const (
-	expHour   = 72
-	secretKey = "secret"
-	UserKey   = "userID"
+	UserKey = "userID"
 )
+
+var (
+	conf authConfig
+)
+
+func Init(accessTokenConf, refreshTokenConf config.AuthToken) {
+	conf = authConfig{
+		AccessToken:  accessTokenConf,
+		RefreshToken: refreshTokenConf,
+	}
+}
 
 func CreateAccessToken(id int) string {
 	claims := jwt.MapClaims{
 		"user_id": strconv.Itoa(id),
-		"exp":     time.Now().Add(time.Hour * expHour).Unix(),
+		"exp":     time.Now().Add(conf.AccessToken.Expires).Unix(),
 	}
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
 
-	tokenString, _ := token.SignedString([]byte(secretKey))
+	tokenString, _ := token.SignedString([]byte(conf.AccessToken.SecretKey))
 
 	return tokenString
 }
@@ -33,22 +41,11 @@ func verifyToken(tokenString string) (*jwt.Token, error) {
 		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
 			return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
 		}
-		return []byte(secretKey), nil
+		return []byte(conf.AccessToken.SecretKey), nil
 	})
 	if err != nil {
 		return nil, err
 	}
 
 	return token, nil
-}
-
-func getTokenFromHeader(r *http.Request) (*jwt.Token, error) {
-	token := r.Header.Get("Authorization")
-	token = strings.TrimPrefix(token, "Bearer ")
-
-	jwtToken, err := verifyToken(token)
-	if err != nil {
-		return nil, err
-	}
-	return jwtToken, nil
 }
