+++ backend/interfaces/api/server/auth/utils.go
@@ -1,6 +1,10 @@
 package auth
 
 import (
+	"net/http"
+	"strings"
+
+	"github.com/dgrijalva/jwt-go"
 	"github.com/google/uuid"
 	"golang.org/x/crypto/bcrypt"
 )
@@ -21,4 +25,15 @@ func IsVerifyPassword(textConplainPassword, hashedPassword string) bool {
 
 func GenerateToken() string {
 	return uuid.Must(uuid.NewRandom()).String()
+}
+
+func getTokenFromHeader(r *http.Request) (*jwt.Token, error) {
+	token := r.Header.Get("Authorization")
+	token = strings.TrimPrefix(token, "Bearer ")
+
+	jwtToken, err := verifyToken(token)
+	if err != nil {
+		return nil, err
+	}
+	return jwtToken, nil
 }
